Move bundle build to common konflux pipeline
This removes the `pipelineSpec` in favor of a `pipelineRef`, which will
remove the need for weekly konflux update chore PRs. This also re-orders
the params in the PipelineRuns, to match the order in the common
pipeline.

The common pipeline also uses OCI artifacts for storage, as opposed to a
Tekton Workspace.

Refs:
 - https://issues.redhat.com/browse/ACM-23301

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>