Use defaults instead of pointers for string fields
The most notable string fields that were using pointers were the Mode
fields, which should be either Enabled or Disabled. The documented
default values are now marked in the CRD, and so Kubernetes should set
them when running in a real cluster. Some local tests needed a crafted
"default" gatekeeper with those fields set, but it just matches what
would be rerturned in a real cluster.

This should allow us to not worry as much about what the default
behavior is for each field: using `IsEnabled` should just work.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>